More changes

+++ onnxruntime/python/_pybind_state.py
@@ -13,9 +13,12 @@ try:
 except ImportError as e:
     warnings.warn("Cannot load onnxruntime.capi. Error: '{0}'.".format(str(e)))
     
-	# If on Windows, check if this import error is caused by the user not installing the 2019 VC Runtime
-    if platform.system().lower() == 'windows':
-        # The VC Redist installer usually puts the VC Runtime dlls in the System32 folder
-        if not os.path.isfile('c:\\Windows\\System32\\vcruntime140_1.dll'):
+    # If on Windows, check if this import error is caused by the user not installing the 2019 VC Runtime
+    # The VC Redist installer usually puts the VC Runtime dlls in the System32 folder
+    # This may not always paint the true picture as anyone building from source using VS 2017 might hit this error
+   	# because the machine might be missing the 2019 VC Runtime but it is not actually needed in that case and the 
+    # import error might actually be due to some other reason.
+    # TODO: Guard against False Positive error message
+    if platform.system().lower() == 'windows' and not os.path.isfile('c:\\Windows\\System32\\vcruntime140_1.dll'):
             warnings.warn("Please install the Visual C++ 2019 runtime and then try again")
